implement permission flow for near you
