+++ app/components/ui/card/styles.tsx
@@ -1,0 +1,20 @@
+import { tva } from "@gluestack-ui/nativewind-utils/tva";
+import { isWeb } from "@gluestack-ui/nativewind-utils/IsWeb";
+const baseStyle = isWeb ? "flex flex-col relative z-0" : "";
+
+export const cardStyle = tva({
+  base: baseStyle,
+  variants: {
+    size: {
+      sm: "p-3 rounded",
+      md: "p-4 rounded-md",
+      lg: "p-6 rounded-xl",
+    },
+    variant: {
+      elevated: "bg-background-0",
+      outline: "border border-outline-200 ",
+      ghost: "rounded-none",
+      filled: "bg-background-50",
+    },
+  },
+});
